fix: corrige errores de compilación en motor contextual
- ContextIntent.kt: reemplaza 'keywords' (inaccesible desde companion
  object) por ContextIntentKind.entries.flatMap
- ContextCompatibilityRegistry.kt: usa string literal
  'enabled_notification_listeners' en lugar de constante eliminada en SDK 36
- ContextIntentEngine.kt: corrige tipo Int vs Long en cálculo de días
- Build previewSafeDebug exitoso

+++ app/src/main/java/com/ordia/app/context/ContextIntentEngine.kt
@@ -431,7 +431,7 @@ object ContextIntentEngine {
             for ((name, day) in dayMap) {
                 if (lower.contains(name)) {
                     val daysUntil = (day.value - today.dayOfWeek.value + 7) % 7
-                    targetDate = if (daysUntil == 0L) today.plusDays(7) else today.plusDays(daysUntil)
+                    targetDate = if (daysUntil == 0) today.plusDays(7) else today.plusDays(daysUntil.toLong())
                     break
                 }
             }
